refactor: BadgeLabel 을 더 제너럴하게 리펙토링
