+++ ios/IssueTracker/IssueTracker/Presentation/Tabs/Issue/View/IssueListCell.swift
@@ -57,9 +57,9 @@ final class IssueListCell: UITableViewCell {
         return label
     }()
 
-    private let badgeLabel: BadgeLabel = {
-        let badgeLabel = BadgeLabel(frame: .zero)
-        badgeLabel.padding = UIEdgeInsets(top: 10, left: 15, bottom: 10, right: 15)
+    private let badgeLabel: Label = {
+        let badgeLabel = Label(frame: .zero)
+		badgeLabel.setPadding(top: 10, left: 15, bottom: 10, right: 15)
         badgeLabel.font = .systemFont(ofSize: 12, weight: .bold)
         badgeLabel.textColor = .white
         badgeLabel.clipsToBounds = true
